# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Test challange",
        "title": "Subscriptions service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/subscription/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "record creation",
                "parameters": [
                    {
                        "description": "item to add",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/subscriptions.Subscription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "created",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscription/delete": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "record deleting",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "record id",
                        "name": "rowId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data deleted",
                        "schema": {
                            "$ref": "#/definitions/subscriptions.Subscription"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscription/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "list of records",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "loaded successfully",
                        "schema": {
                            "$ref": "#/definitions/subscriptions.SubscriptionListPage"
                        }
                    },
                    "405": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscription/read": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "record reading",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "record id",
                        "name": "rowId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data found",
                        "schema": {
                            "$ref": "#/definitions/subscriptions.Subscription"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscription/sum": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "sum calculation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "period from",
                        "name": "filterFrom",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "period to",
                        "name": "filterTo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sum is ready",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscription/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "record update",
                "parameters": [
                    {
                        "description": "item to update",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/subscriptions.Subscription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "subscriptions.Subscription": {
            "type": "object",
            "properties": {
                "finish_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "subscriptions.SubscriptionListPage": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/subscriptions.Subscription"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}